nbg6817-dualboot: add --get-bootloader-rootfs
Provide an alternative means to determine the rootfs configured by the
bootloader, without reading /proc/cmdline (which is used by sysupgrade
and might be overmounted in order to write to the current partition set
instead of of the alternative one). This parameter reads the bootloader
value from /sys/firmware/devicetree/base/chosen/bootloader-args.

Signed-off-by: Stefan Lippers-Hollmann <s.l-h@gmx.de>